backend-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2
websockets==13.1
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
